refactor: remove settings.json and update component imports for consistency
