+++ frontend/src/components/common/RLogotipo.vue
